+++ app/helpers/custom_fields_helper.rb
@@ -13,7 +13,7 @@
 module CustomFieldsHelper
 
   def custom_fields_tabs
-    tabs = [{:name => 'WorkPackageCustomField', :partial => 'custom_fields/index', :label => :label_issue_plural},
+    tabs = [{:name => 'WorkPackageCustomField', :partial => 'custom_fields/index', :label => :label_work_package_plural},
             {:name => 'TimeEntryCustomField', :partial => 'custom_fields/index', :label => :label_spent_time},
             {:name => 'ProjectCustomField', :partial => 'custom_fields/index', :label => :label_project_plural},
             {:name => 'VersionCustomField', :partial => 'custom_fields/index', :label => :label_version_plural},