+++ doc/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+* `#1724` PDF Export of Work Packages with Description
+
 ## 3.0.0pre12
 
 * `#1417` Enable default behavior for types