+++ doc/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 * `#1724` PDF Export of Work Packages with Description
+* `#1731` Squash old migrations into one
 
 ## 3.0.0pre12