+++ doc/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 * `#1421` Adapt issue created/updated wording to apply to work packages
+* `#1813` Run Data Generator on old AAJ schema
 
 ## 3.0.0pre13