+++ doc/CHANGELOG.md
@@ -1,9 +1,10 @@
 # Changelog
 
-* `#825` Migrate Duration
 * `#1421` Adapt issue created/updated wording to apply to work packages
 * `#1768` Fixed bug: Klicking on Wiki Edit Activity results in error 500
 * `#1813` Run Data Generator on old AAJ schema
+* `#1859` Fix 20130814130142 down-migration (remove_documents)
+* `#825`  Migrate Duration
 
 ## 3.0.0pre13