+++ doc/CHANGELOG.md
@@ -3,11 +3,11 @@
 * `#825` Migrate Duration
 * `#828` Remove Alternate Dates
 * `#1421` Adapt issue created/updated wording to apply to work packages
+* `#1610` Move Planning Element Controller to API V2
 * `#1768` Fixed bug: Klicking on Wiki Edit Activity results in error 500
 * `#1787` Remove Scenarios
 * `#1813` Run Data Generator on old AAJ schema
 * `#1859` Fix 20130814130142 down-migration (remove_documents)
-* `#825`  Migrate Duration
 
 ## 3.0.0pre13
 
@@ -23,6 +23,7 @@
 
 * `#1417` Enable default behavior for types
 * `#1631` Remove documents from core
+* `#1686` Issues not accessible in public projects when not a member
 
 ## 3.0.0pre11